+++ CHANGELOG.md
@@ -754,3 +754,4 @@
 * Move `Interpreted` module to `Core`
 * Move `Schematic` type family to `Schemes` module
 * Move `Vertical` and `Horizontal` datatypes to `Auxiliary` module
+* Rename `Horizontal` constructors to `Left` and `Right`